backend/pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "ubersicht-spotify-plugin"
version = "0.0.1"
description = ""
license = "MIT"
authors = [
    "Igor Lashkov <rwrotson@yandex.ru>",
]
readme = "../README.md"
packages = [{include = "app", from = "src"}]
repository = "https://github.com/rwrotson/ubersicht-spotify-plugin"
keywords = [
    "music",
    "plugin",
    "ubersicht",
    "spotify",
]

[tool.poetry.dependencies]
python = "3.12.1"
fastapi = "0.106.0"
uvicorn = {extras = ["standard"], version = "0.25.0"}
python-dotenv = "1.0.0"
aiohttp = "3.9.1"


[tool.poetry.group.dev.dependencies]
pytest = "7.4.3"
pytest-cov = "4.1.0"
mypy = "1.6.1"
flake8 = "6.1.0"
tox = "4.11.3"
black = "23.10.1"
isort = "5.12.0"

[tool.pytest.ini_options]
addopts = "--cov=app"
testpaths = [
    "tests",
]

[tool.mypy]
mypy_path = "src"
check_untyped_defs = true
disallow_any_generics = true
ignore_missing_imports = true
no_implicit_optional = true
show_error_codes = true
strict_equality = true
warn_redundant_casts = true
warn_return_any = true
warn_unreachable = true
warn_unused_configs = true
no_implicit_reexport = true

[tool.poetry.scripts]
start = "app.main:start"